+++ client/js/middleware/serialization/qbank/deserializers/base.js
@@ -3,14 +3,19 @@ import { getQbankType }     from '../../../../constants/genus_types';
 import { getLanguage }      from '../../../../constants/language_types';
 
 export default function base(item) {
-  debugger
+  const languages = item.question ? item.question.texts : [];
+  let languageTypeId = '';
+  // TODO: always grabs last one, pass the langauge through and find it.
+  if (!_.isEmpty(languages)) {
+    languageTypeId = _.last(languages).languageTypeId;
+  }
   return {
     id: item.id,
     type: getQbankType(item.genusTypeId),
     bankId: item.bankId,
     assessmentId: null, // TODO
     name: _.get(item, 'displayName.text'),
-    language: getLanguage(_.get(item, 'question.text.languageTypeId')), // TODO: How does the language even work?
+    language: getLanguage(languageTypeId), // TODO: How does the language even work?
     question: {
       id: _.get(item, 'question.id'),
       type: getQbankType(_.get(item, 'question.genusTypeId')),
