+++ client/js/middleware/serialization/qbank/serializers/base.js
@@ -1,12 +1,40 @@
 import _                from 'lodash';
 import genusTypes       from '../../../../constants/genus_types';
 import { scrub }        from '../../serializer_utils';
+import { languages, getLanguage } from '../../../../constants/language_types';
 
-export function baseSerializeQuestion(originalItem, newAttributes) {
+function serializeQuestionString(item) {
+  if (item.name) {
+    // create
+    const simpleLanguage = getLanguage(item.language);
+    item.question = {
+      text: item.question.text,
+      languageTypeId: item.language,
+      formatTypeId: languages.formatTypeId,
+      scriptTypeId: languages.scriptTypeId[simpleLanguage]
+    };
+    return item.question;
+  } else if (item.language) {
+    // update language
+    const simpleLanguage = getLanguage(item.language);
+    item.question = {
+      text: '',
+      languageTypeId: item.language,
+      formatTypeId: languages.formatTypeId,
+      scriptTypeId: languages.scriptTypeId[simpleLanguage]
+    };
+    return item.question;
+  } else {
+    // update
+    return item.question.text;
+  }
+}
+
+export function baseSerializeQuestion(originalItem, newAttributes, item) {
   return {
     id: _.get(originalItem, 'question.id'),
     genusTypeId: genusTypes.question[_.get(newAttributes, 'type') || originalItem.type],
-    questionString: newAttributes,
+    questionString: serializeQuestionString(item),
     fileIds: newAttributes.fileIds
   };
 }
@@ -17,7 +45,7 @@ export function baseSerializeItem(originalItem, newAttributes) {
     genusTypeId: genusTypes.item[_.get(newAttributes, 'type') || originalItem.type],
     name: _.get(newAttributes, 'name', originalItem.name),
     question: scrub(
-      baseSerializeQuestion(originalItem, _.get(newAttributes, 'question', {}))
+      baseSerializeQuestion(originalItem, _.get(newAttributes, 'question', {}), newAttributes)
     ),
     answers: null,
   };
